update shopping cart

+++ commerce/interaction-api/src/main/java/ru/yandex/practicum/commerce/dto/shopping/cart/ChangeProductQuantityRequestDto.java
@@ -22,5 +22,5 @@ public class ChangeProductQuantityRequestDto {
 
     @NotNull
     @PositiveOrZero
-    int newQuantity;
+    Long newQuantity;
 }
